chore(refactor): jouney token middleware
Make use of a middleware to get the journey auth token.

You can finde more info on middlewares here:
https://expressjs.com/en/guide/writing-middleware.html
