+++ src/middleware/journey-token.js
@@ -1,11 +1,20 @@
 const request = require('request')
 
+module.exports = function journeyToken (req, res, next) {
+  getToken((err, token) => {
+    if (err) return next(err)
+    req.journeyToken = token
+    next()
+  })
+}
+
 const getTokenData = {
   client_secret: process.env.JOURNEY_TOKEN_CLIENT_SECRET,
   client_id: process.env.JOURNEY_TOKEN_CLIENT_ID,
   grant_type: 'client_credentials'
 }
-module.exports = function getToken (cb) {
+
+function getToken (cb) {
   request.post(
     process.env.JOURNEY_TOKEN_URL,
     {
